Added JavaScript code
Added JavaScript code to validate the form and an alert message for the form validator. Added a dropdown menu for the gallery section in the navigation area. This makes it so the user can access the form to share their experience without having to go through the gallery page. It would have not made a lot of sense for me to add an image slideshow because the gallery would hypothetically be an unidentified amount of images and I could not imagine how to do that with a slideshow in JavaScript.